fix: use float64 for extra_usage fields to match API response
The usage API returns used_credits and monthly_limit as floats (e.g.
0.0), but the struct used *int, causing every successful response to
fail JSON unmarshalling. This was cached as a failure with only 15s
TTL, triggering a new API call every 15 seconds until rate limited.

+++ main.go
@@ -86,9 +86,9 @@ type quotaLimit struct {
 
 // extraUsage is the pay-as-you-go overage info.
 type extraUsage struct {
-	IsEnabled    bool `json:"is_enabled"`
-	MonthlyLimit *int `json:"monthly_limit"`
-	UsedCredits  *int `json:"used_credits"`
+	IsEnabled    bool     `json:"is_enabled"`
+	MonthlyLimit *float64 `json:"monthly_limit"`
+	UsedCredits  *float64 `json:"used_credits"`
 }
 
 // usageResponse is the API response from the usage endpoint.
@@ -438,8 +438,8 @@ func formatExtraUsage(extra *extraUsage) string {
 	if extra == nil || !extra.IsEnabled || extra.MonthlyLimit == nil || extra.UsedCredits == nil {
 		return ""
 	}
-	used := *extra.UsedCredits / 100
-	limit := *extra.MonthlyLimit / 100
+	used := int(*extra.UsedCredits) / 100
+	limit := int(*extra.MonthlyLimit) / 100
 	return fmt.Sprintf("$%d/$%d", used, limit)
 }
